Add files via upload
A peer-to-peer (P2P) website is a platform that facilitates direct communication and exchange of information or resources between users, without the need for intermediaries or centralized control. The purpose of this project is to develop a peer-to-peer website that enables users to interact, share content, and collaborate directly with each other.